Relax the requirements for the Recipient field
PHP is seemingly sending the recipient in the wrong format and causing
Jetpack to fail to send email on 16.04. This appears to be due to the
malformed "To:" field.

From an `strace` of the PHP process, it looks like this is the contents
of the "To:" field:

    To: \"\\\"michael.herold@getflywheel.com\\\"\" <michael.herold@getflywheel.com>

Clearly, the escaping is wrong, but the actual email address comes
through okay. Thus, we should be able to do a second pass from the start
of the string to check if there is the less-than sign to grab the email
address from the `angle-addr` (see [RFC5322 Section 3.4] for the
definition of `angle-addr`) section of the recipient.

This relaxation of the requirements for the field should let us zero in
on the `angle-addr` to eliminate this problem.

I have so far tested this manually on the following recipients:

    \"\\\"michael.herold@getflywheel.com\\\"\" <michael.herold@getflywheel.com>
    <michael.herold@getflywheel.com>
    michael.herold@getflywheel.com

And they all seem to pass. I'm going to work on some tests so we can
check for regressions if we need to make any further changes.

[RFC5322 Section 3.4]: https://tools.ietf.org/html/rfc5322#section-3.4